Add RemoteSource dummy and complete RepoDi

+++ app/src/main/java/sidev/app/bangkit/capstone/sheltermobile/core/domain/model/WeatherForecast.kt
@@ -3,7 +3,7 @@ package sidev.app.bangkit.capstone.sheltermobile.core.domain.model
 import java.sql.Timestamp
 
 data class WeatherForecast(
-    val temperature: Double,
+    val temperature: Float,
     val humidity: Float,
     val rainfall: Float,
     val windSpeed: Float,
